Cleanup
- Restructure Billing directory
- Delete leftover unused code from original version
