+++ packages/manager/src/features/Billing/BillingDetail.tsx
@@ -11,9 +11,9 @@ import { AccountsAndPasswords, BillingAndPayments } from 'src/documentation';
 import { useAccount } from 'src/hooks/useAccount';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import BillingActivityPanel from './BillingPanels/BillingActivityPanel';
-import BillingSummary from './BillingSummary';
-import ContactInfo from './BillingPanels/SummaryPanel/PanelCards/ContactInformation';
-import PaymentInformation from './BillingPanels/SummaryPanel/PanelCards/PaymentInformation';
+import BillingSummary from './BillingPanels/BillingSummary';
+import ContactInfo from './BillingPanels/ContactInfoPanel';
+import PaymentInformation from './BillingPanels/PaymentInfoPanel';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {},
