+++ packages/manager/src/features/Billing/BillingPanels/BillingActivityPanel/BillingActivityPanel.tsx
@@ -36,7 +36,7 @@ import { useSet } from 'src/hooks/useSet';
 import { isAfter } from 'src/utilities/date';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import { getAll, getAllWithArguments } from 'src/utilities/getAll';
-import { getTaxID } from '../RecentPaymentsPanel/RecentPaymentsPanel';
+import { getTaxID } from '../../billingUtils';
 
 const useStyles = makeStyles((theme: Theme) => ({
   headerContainer: {
