+++ packages/manager/src/features/Billing/BillingPanels/ContactInfoPanel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ContactInformation';
