+++ packages/manager/src/features/Billing/BillingPanels/PaymentInfoPanel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PaymentInformation';
