--- packages/manager/src/features/Billing/BillingPanels/PromotionsPanel/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './PromotionsPanel';
