--- packages/manager/src/features/Billing/BillingPanels/RecentPaymentsPanel/index.ts
@@ -1,2 +1,0 @@
-import RecentPaymentsPanel from './RecentPaymentsPanel';
-export default RecentPaymentsPanel;
