--- packages/manager/src/features/Billing/BillingPanels/SummaryPanel/index.tsx
@@ -1,2 +1,0 @@
-import SummaryPanel from './SummaryPanel';
-export default SummaryPanel;
