--- packages/manager/src/features/Billing/BillingPanels/UpdateCreditCardPanel/index.ts
@@ -1,2 +1,0 @@
-import UpdateCreditCardPanel from './UpdateCreditCardPanel';
-export default UpdateCreditCardPanel;
