+++ packages/manager/src/features/Billing/billingUtils.ts
@@ -6,3 +6,16 @@ export const cleanCVV = (input: string): string => {
   const cvv = input.slice(0, 4);
   return cvv.replace(regex, '');
 };
+
+export const getTaxID = (
+  invoiceItemDate: string,
+  taxDate?: string,
+  taxID?: string
+) => {
+  if (!taxID || !taxDate) {
+    return undefined;
+  }
+  const taxStartedBeforeThisInvoiceItem =
+    Date.parse(invoiceItemDate) > Date.parse(taxDate);
+  return taxStartedBeforeThisInvoiceItem ? taxID : undefined;
+};
